Time: 58 ms (5.93%), Space: 9.5 MB (6.42%) - LeetHub

+++ 2083-substrings-that-begin-and-end-with-the-same-letter/2083-substrings-that-begin-and-end-with-the-same-letter.cpp
@@ -3,13 +3,13 @@
 class Solution {
 public:
     long long numberOfSubstrings(string s) {
-        unordered_map<char, long long> degree;
+        unordered_map<char, int> degree;
         for (auto ch : s) {
             degree[ch]++;
         }
         long long ans = 0;
         for (auto it : degree) {
-            ans += (it.second*(it.second + 1ll))/2ll;
+            ans += ((long long)it.second*(it.second + 1ll))/2ll;
         }
         return ans;
     }
